% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitLimit.R
\name{twitLimit}
\alias{twitLimit}
\title{Twitter Request Limit}
\usage{
twitLimit(twitter_token)
}
\arguments{
\item{twitter_token}{Your twitter access token.}
}
\description{
Retrieves number of requests remaining from your twitter token.
Requests are renewed every-so-often (typically 15 minutes) based on Twitter's API user agreement.
}
\examples{
twitter_token <- twitToken(consumer_key, consumer_secret, access_token, access_secret)
limit <- twitLimit(twitter_token, query = 'search')
}

