% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitS1.R
\name{twitS1}
\alias{twitS1}
\title{Request Twitter User Timeline Once}
\usage{
twitS1(twitter_token, query)
}
\arguments{
\item{twitter_token}{Your twitter_token; e.g. from twitToken() function.}

\item{query}{A user_timeline query; e.g. from twitUTQuery() function.}
}
\description{
Make a single request to extract tweets with query specifications.
}
\examples{
twitter_token <- twitToken(consumer_key, consumer_secret, access_token, access_secret)
query <- twitSQuery(screen_name = 'thomas_laetsch')
user.data <- twitS1(twitter_token, query)
user.json.data <- httr::content(user.data)
}

