% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitSQuery.R
\name{twitSQuery}
\alias{twitSQuery}
\title{Twitter Search Tweets Query Maker}
\usage{
twitSQuery(q = "", count = 100, geocode = NULL, lang = NULL,
  result_type = NULL, until = NULL, since_id = NULL, max_id = NULL)
}
\arguments{
\item{q}{Character. Keyword query to search tweets.}

\item{count}{(defaults to 100) Integer. Number of tweets to return per request; max = 100.}

\item{geocode}{(optional) Character. Pull tweets via geolocation within given search radius.}

\item{result_type}{(optional) Character. Pull tweets which are recent, popular, or mixed.}

\item{until}{(optional) Character. Pull tweets appearing before given date.}

\item{since_id}{(optional) Integer. Request results with tweet id greater (newer) than this value.}

\item{max_id}{(optional) Integer. Request results with tweet id less (older) than or equal to this value.}

\item{language}{(optional) Character. Pull tweets of a specific language.}
}
\description{
Sets up URL encoded query list to query for user timelines in the statuses/user_timeline request.
Most commonly use as support function for twitUT1.
}
\examples{
twitter_token <- twitToken(consumer_key, consumer_secret, access_token, access_secret)
query <- twitSQuery(q = 'thomas laetsch ua OR ucsd OR uconn OR nyu')
user.data <- twitS1(twitter_token, query)
user.json.data <- httr::content(user.data)
}
\seealso{
\link{https://dev.twitter.com/rest/public/search}
}

