% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitUT1.R
\name{twitUT1}
\alias{twitUT1}
\title{Request Twitter User Timeline Once}
\usage{
twitUT1(twitter_token, query)
}
\arguments{
\item{twitter_token}{Your twitter_token; e.g. from twitToken() function.}

\item{query}{A user_timeline query; e.g. from twitUTQuery() function.}
}
\description{
Make a single request to extract user timeline tweets with query specifications.
}
\examples{
twitter_token <- twitToken(consumer_key, consumer_secret, access_token, access_secret)
query <- twitUTQuery(screen_name = 'thomas_laetsch')
user.data <- twitUT1(twitter_token, query)
user.json.data <- httr::content(user.data)
}

