% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitUTmany2json.R
\name{twitUTmany2json}
\alias{twitUTmany2json}
\title{Write many Twitter User Timeline requests to JSON file}
\usage{
twitUTmany2json(filename, twitter_token, n = 5000, screen_name = NULL,
  user_id = NULL, since_id = NULL, max_id = NULL, trim_user = "true",
  exclude_replies = NULL, contributor_details = NULL, verbose = TRUE,
  append_old = FALSE)
}
\arguments{
\item{filename}{Character. File path to json file to store data.}

\item{twitter_token}{Your twitter_token; e.g. from twitToken() function.}

\item{n}{(optional) Integer. Maximum total tweets to pull from the timeline.}

\item{screen_name, }{user_id, since_id, max_id, trim_user, exclude_replies, contributor_details 
To set up query. See ?twitUTQuery for details}

\item{verbose}{(defaults TRUE) Logical. If true, displays request limit at each request.}

\item{append_old}{(defaults FALSE) Logical. If true will append an already extant json file; otherwise it creates/overwrites file.}
}
\description{
Used to extract many repeated requests to pull a user's timeline and save the results to a json file.
Will automatically move from most recent toward older tweets while pulling data.
}
\examples{
twitter_token <- twitToken(consumer_key, consumer_secret, access_token, access_secret)
twitUTmany2json(filename='~/thomas_laetsch.json', twitter_token = twitter_token, screen_name = 'thomas_laetsch')
}

