% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitUTQuery.R
\name{twitUTQuery}
\alias{twitUTQuery}
\title{Twitter User Timeline Query Maker}
\usage{
twitUTQuery(screen_name = NULL, user_id = NULL, count = 200,
  since_id = NULL, max_id = NULL, trim_user = "true",
  exclude_replies = NULL, contributor_details = NULL)
}
\arguments{
\item{screen_name}{(optional if user_id is specified) Character. Screen name of user to request the timeline from.}

\item{user_id}{(optional if screen_name is specified) Integer. User ID of user to request the timeline from.}

\item{count}{(optional, default = 200) Integer. Number of tweets per request; max = 200.}

\item{since_id}{(optional) Integer. Request results with tweet id greater (newer) than this value.}

\item{max_id}{(optional) Integer. Request results with tweet id less (older) than or equal to this value.}

\item{trim_user}{(optional) Logical. If true returns only user_id in user object. False returns complete user object.}

\item{exclude_replies}{(optional) Logical. If true prevents replies on user timeline from being returned.}

\item{contributor_details}{(optional) Logical. If true, contributor's screen_name is also returned, instead of only user_id.}
}
\description{
Sets up URL encoded query list to query for user timelines in the statuses/user_timeline request.
Most commonly use as support function for twitUT1.
}
\examples{
twitter_token <- twitToken(consumer_key, consumer_secret, access_token, access_secret)
query <- twitUTQuery(screen_name = 'thomas_laetsch')
user.data <- twitUT1(twitter_token, query)
user.json.data <- httr::content(user.data)
}
\seealso{
\link{https://dev.twitter.com/rest/reference/get/statuses/user_timeline}
}

